requirements.txt
boto3==1.34.134
mysql-connector-python==9.0.0
matplotlib==3.9.2